Rename ExitPopup to EindePopup
